+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+## vNext
+
+### Features
+
+- Add Bubble component [PR 31](https://github.com/input-output-hk/react-polymorph/pull/31)
+- Use Bubble component in Options component [PR 31](https://github.com/input-output-hk/react-polymorph/pull/31)
+
+### Fixes
+
+- Fix bug causing crash when hitting backspace in an Autocomplete component [PR 31](https://github.com/input-output-hk/react-polymorph/pull/31)
+
 ## 0.5.5
 
 ### Fixes